V1.1.6.5 Fix a bug that will display negative of smallint data type when value >= 128 and <= 255.

+++ C#/src/Hubble.Data/Hubble.Core/Data/DBProvider.cs
@@ -1380,7 +1380,7 @@ namespace Hubble.Core.Data
                     return (long)(((sbyte*)(&payloadData[_DocIdReplaceField.TabIndex]))[_DocIdReplaceField.SubTabIndex]);
                 case DataType.SmallInt:
                     sbyte* sbyteP = (sbyte*)&payloadData[_DocIdReplaceField.TabIndex] + _DocIdReplaceField.SubTabIndex;
-                    short l = *sbyteP;
+                    short l = (byte)(*sbyteP);
                     short h = *(sbyteP + 1);
                     h <<= 8;
                     return l | h;
